working fix

+++ server/routes/proxyroute.js
@@ -17,20 +17,35 @@ router.delete("/proxies", async function (req, res) {
     try {
         var apiKey = req.body.apiKey;
         var userId = req.body.userId;
+
+        if (typeof userId === "undefined" || typeof apiKey === "undefined") {
+            res.send(400, "Missing body elements");
+            return;
+        }
+
         var x = new proxy(apiKey);
+
+        // Step 1. Validate API token
+        await x.validateToken(apiKey);
 
         var proxySchema = require('./proxySchema');
         var response = await proxySchema.find({
             'userId': userId
         });
-        var response2 = await proxySchema.deleteMany({
-            'userId':userId
-        });
 
-        for (var i = 0; i < response2.n; ++i) {
-            x.deleteInstance(response[i].instanceId);
+        if (response.length > 0) {
+            // Step 2. Delete all proxies from servers
+            await x.deleteAllInstances(apiKey);
+
+            // Step 3. Delete all proxies from database
+            await proxySchema.deleteMany({
+                'userId': userId
+            });
+
+            res.send(200, "Delete successful.");
+        } else {
+            res.send(400, "userId, not found");
         }
-        res.statusCode(200).send(response2.n);
     } catch (err) {
         res.send(err.statusCode);
     }
@@ -41,18 +56,30 @@ router.delete("/proxy", async function (req, res) {
     try {
         var apiKey = req.body.apiKey;
         var proxyId = req.body.proxy;
-        var x = new proxy(apiKey);
 
-        var proxySchema = require('./proxySchema');
-        var response = await proxySchema.findOneAndRemove({
-            'proxy': proxyId
-        });
-        var result = await x.deleteInstance(response.instanceId)
-        res.send(result);
+        if (typeof proxyId === "undefined" || typeof apiKey === "undefined") {
+            res.send(400, "Missing body elements");
+        } else {
+            var x = new proxy(apiKey);
+
+            var proxySchema = require('./proxySchema');
+            var response = await proxySchema.findOne({
+                'proxy': proxyId
+            });
+
+            if (response != null) {
+                var result = await x.deleteInstance(apiKey, response.instanceId)
+                await response.delete();
+                res.send(result);
+            }
+            res.send(400, "Not found");
+        }
+
     } catch (err) {
         res.send(err.statusCode);
     }
 });
+
 
 // Get all regions for linode provider
 router.get("/regions", async function (req, res) {
@@ -76,7 +103,7 @@ router.get("/proxies", async function (req, res) {
             'userId': userId
         });
         console.log(`Fetching proxies for ${userId}`)
-        query.select('proxy region instanceId userId');
+        query.select('proxy region instanceId userId, server');
         query.exec(function (err, result) {
             res.send(result);
         });
@@ -87,7 +114,6 @@ router.get("/proxies", async function (req, res) {
 
 // Create proxies for linode provider
 router.post("/proxies", async function (req, res) {
-
     try {
         var apiKey = req.body.apiKey;
         var region = req.body.region;
@@ -100,16 +126,57 @@ router.post("/proxies", async function (req, res) {
         var result = await x.generateProxies(userId, number, user, pass, region);
         res.send(result);
     } catch (err) {
-
         console.log(err);
         if (err instanceof RangeError) {
             console.log('No servers to generate proxies from');
         }
-        else {
+        else if (err.statusCode) {
             res.send(err.statusCode);
+        } else {
+            console.log(err);
         }
     }
 
 });
 
+router.post("/exec", async function (req, res) {
+    try {
+
+        node_ssh = require('node-ssh');
+        ssh = new node_ssh();
+
+        await ssh.connect({
+            'host':'172.104.26.36',
+            'username':'root',
+            'password': '5fgdj423fds'
+        });
+
+        var bashCommand =
+        "yum -y update && " +
+        "yum -y install squid httpd-tools wget && " +
+        "systemctl start squid && " +
+        "systemctl enable squid && " +
+        "touch /etc/squid/passwd && " +
+        `htpasswd -b /etc/squid/passwd profileaio 5fgdj423fds && ` +
+        "wget -O /etc/squid/squid.conf https://raw.githubusercontent.com/dzt/easy-proxy/master/confg/userpass/squid.conf && " +
+        "systemctl restart squid.service && " +
+        "systemctl enable squid.service && " +
+        `iptables -I INPUT -p tcp --dport 3128 -j ACCEPT && ` +
+        "iptables-save";
+
+        await ssh.execCommand(bashCommand).then(function(result) {
+            console.log('STDOUT: ' + result.stdout)
+            console.log('STDERR: ' + result.stderr)
+        });
+
+
+    } catch (Err) {
+        console.log(Err);
+    }
+
+});
+
 module.exports = router;
+
+
+
